requirements.txt
streamlit==1.32.0
requests==2.32.3
python-dotenv==1.0.1
SpeechRecognition==3.10.1
pyaudio==0.2.14
numpy==1.26.4
elevenlabs==0.3.0
soundfile==0.12.1
pydub==0.25.1